- Fixing server.dao.test - Fixing server.js for testing

+++ pulsebs-server/pulsebsDAO.js
@@ -38,24 +38,26 @@ function loadData() {
 /*
 * Database Connection
 */
-if ( process.env.TEST && process.env.TEST === '1' ) {
-    //TEST DB
-    const dbname = 'pulsebs-test.db'
-    //Check for file existance
-    if ( fs.existsSync( dbname ) ) {
-        fs.unlinkSync( dbname ); //If true, delete file
+if( !db ) {
+    if ( process.env.TEST && process.env.TEST === '1' ) {
+        //TEST DB
+        const dbname = 'pulsebs-test.db'
+        //Check for file existance
+        if ( fs.existsSync( dbname ) ) {
+            fs.unlinkSync( dbname ); //If true, delete file
+        }
+        //Create new db
+        db = new sqlite3.Database( 'pulsebs-test.db', ( err ) => {
+            if ( err ) return console.error( err.message );
+            else console.log( 'Connected to the in-memory TEST SQlite database.' );
+        } );
+        loadData();
+    } else {
+        db = new sqlite3.Database( 'pulsebs.db', ( err ) => {
+            if ( err ) return console.error( err.message );
+            else console.log( 'Connected to the in-memory SQlite database.' );
+        } );
     }
-    //Create new db
-    db = new sqlite3.Database( 'pulsebs-test.db', ( err ) => {
-        if ( err ) return console.error( err.message );
-        else console.log( 'Connected to the in-memory TEST SQlite database.' );
-    } );
-    loadData();
-} else {
-    db = new sqlite3.Database( 'pulsebs.db', ( err ) => {
-        if ( err ) return console.error( err.message );
-        else console.log( 'Connected to the in-memory SQlite database.' );
-    } );
 }
